requirements.txt
# Python Requirements
gurux-dlms
beautifulsoup4
lxml
pyserial
influxdb-client
cryptography==37.0.2
